Completely re-work the protocol definitions.
The XML has been re-done without the telepathy extensions.
An 'idl' has been added. This is NOT currently the canonical
protocol definition. It has more information than the XML
spec but is written in a pseudo idl for which there is no
parser. In the future the protocol should be defined to a similar
standard. The requirements for moving to canonical IDL
are that it is useful for generating Qt / GBus bindings.

Add a script to create c literals of introspection
data out of the introspection XML.