Move cache updates to an idle
When receiving a children-changed signal, we add a whole subtree to the
cache, but not all of the objects have necessarily been initialized yet,
so we may, eg, indicate in the cache that an object's role is UNKNOWN
(see BNC#600024).
This also fixes a crash when starting Firefox 3.6 with at-spi2.