enable peer to peer comunication for at-spi2
Each application will setup a dbus server listening on a unix socket in
/tmp/at-spi2/xxxxxx.  A getApplicationBusAddress method is added to
atk-adaptors/adaptors/application-adaptor.c so that clients of at-spi2
can get the address of the socket on which to contact the application
server.  This method returns a string which is a dbus address so it can
be passed directly to dbus_connection_open.