Fix for BGO#643110: Do not allow all users to connect to the accessibility bus
Allowing all users to connect to the accessibility bus creates a security
issue where one user could control applications being run by another user.
So we'll only allow root to connect, so that applications running as root are
still accessible.  This assumes that dbus-daemon is running as the user who
is currently logged in.