Use libatspi  to get the accessibility bus and handle main loop integration
Remove function to get the accessibility bus, since it was duplicating code
in at-spi2-core and should be modified for BGO#644851.
Also, use atspi_dbus_*_setup_with_g_main instead of depending on dbus-glib,
so that FDO#35115 doesn't show up and we don't need to disable p2p if an old
version of dbus-glib is installed.

Note that at-spi2-atk now depends on libatspi where previously it did not.