Call g_error if unable to find the accessibility bus
Previously, AT-SPI would wind up passing a NULL connection to libdbus if it
couldn't connect to the bus, causing a cryptic abort.  This change makes it
at least abort with a sensible error message in that case.
Also, if unable to find the accessibility bus address via the session bus,
save and print the DBusError, if any.