Fix/rework selective caching
Previously, atspi_accessible_set_cache_mask would not affect accessibles
which were already cached, but this was not very useful, as caching is done
automatically when a new app is discovered.  Functions that may retrieve
cached data now test the cache setting for the application--not just whether
there is a value stored for the data--whe ndetermining whether to call the app.

Also, annotate atspi_accessible_set_cache_mask as a gint.  This is temporary;
properly generating GTypes for enums/flags is TODO for 2.2.