Disable caching when not in atspi_event_main
If we are not in a GLib main loop, then we will not see signals to update
the cache, so cached data will become invalid.  Assume that, if
atspi-event_main has not been called, then we are, for instance, in a simple
pyatspi script without a main loop, so we should disable caching.