Remove redundant tables of state and role names
Use the generated enum type to convert state names to numeric values rather
than a string table that needs to be maintained in parallel.  Also, remove
atspi_role_get_name since it was unused, marked deprecated, and relied on a
string table which is now redundant with the enum type.