BGO#652858: Deregister objects on state-changed::defunct
When a state-change:defunct is received, remove the object from the
dbus-path-to-object-mapping hash.  Also, avoid enumerating an object's
children in a call to GetItems if the object is marked defunct.  This
should mostly shield the ATK implementation from being called on objects
that have been marked defunct.

Also removed a seemingly unnecessary g_strdup/g_free pair.