Fix a failure to respond to children-changed events in some cases
When a children-changed event is received, the cache of children should
be updated if present.  However, to test this, the code was checking for
the presence of a non-NULL list of children, which would be the case if
the object initially had 0 children.  The correct thing to do is to
check the accessible's cached_properties.