API changes to device and event listeners for compatibility with js
Skip atspi_device_listener_new_simple and
atspi-event_listener_new_simple, since Javascript throws an error for
callback functions that don't provide user_data.
Add atspi_event_listener_new.
Change order of user_data and destroy notify parameters in
atspi_device_listener_new to match the order used other places in GNOME.
The old order was triggering an IndexError because of what appears to be
a bug in pygobject.  Note that this is an API change and would
ordinarily be a bad idea, but at this point I don't know of anyone using
libatspi directly, so going to commit it.
Some introspection fixes.