Export atk_misc_intance for VS/Win32 builds
The variable atk_misc_intance needs to be explicitly decorated
with the __declspec(dllexport) directive to export properly on Windows/VS
DLL builds (which is the standard build).  The corresponding mechanism
for linking to the ATK DLL is also updated as well with
extern __declspec(dllimport).

This will allow GTK+ 3.1.9+ (with the a11y functions)to link correctly on
Visual C++ .