Using atk_major_version on the binary_age
This was required in order to avoid a shlib downgrade due the
bump to 2.0

The ideal would be to update all the 1.0 stuff to 2.0. But that
would required libraries using atk-1.0 to be modified in order
to link against 2.0. Taking into account that ATK 2.0 is
fully API and ABI compatible with ATK 1.0, this seems an
overkill.