Fix out of tree compilation on Win32
The Makefile.am file contains some Win32-specific instructions.
Two of these instructions contain references to $(srcdir)/atk.def
as the location of the file containing a list of symbols which
need to be exported in the shared library (DLL on Win32).

However, as this file is auto-generated during the build this
file will end up in the build dir and not in the source dir.
For people who don't use out of tree compilation this won't
cause any problems, but for people who are the compilation
will fail as this file doesn't exist in the $(srcdir).

Fixed this by changing these references from $(srcdir)/atk.def to atk.def