Using new format to register to window events
Now ATK has AtkWindow, so the format to register to window events
is like any other ATK event

We check if we need to register using the old format, assuming that
if you are able to register to a window event using the old format
is because the ATK implementor is still using the old one. This
check should be temporal.