Prevent gnome-shell getting stuck on the login screen
The workaround that was committed for this didn't have the
intended effect, since the timeout was added to the default
main context, not the one that is used in the recursive
mainloop.

Without this patch, my login screen would freeze when hitting
Enter in the password entry, with it, I could successfully log
in 10 out of 10 times.

https://bugzilla.gnome.org/show_bug.cgi?id=658013