Support sub-details for events
Firefox sometimes adds a ":system" suffix to events, and event listeners
should automatically catch events with this suffix, as this was done in
AT-SPI1, and the "system" suffix may become a documented part of the ATK
spec (see bug 639466). Note that atk uses a : to delimit the "system"
suffix in the detail, but this is changed to a / over D-Bus to allow use
of arg0path when constructing a match rule.