Avoid making calls on processes thought to be hung
If a method call times out, record the process as being hung, and send a
ping. Consider the process to be no longer hung when a response to the
ping is received.

Hopefully this will prevent long delays if an AT makes consecutive calls
on an application that is not responding.