Invalidate states when receiving a focus event
Sometimes we get "focus" events from gtk, but the cache still does not
indicate that the object has STATE_FOCUSED. Really should get to the
bottom of this but don't have time to investigate more right now. Might
be some kind of issue with the accessible being added to the cache
before it is fully initialized--perhaps AddAccessible really doesn't
make sense.

https://bugzilla.gnome.org/show_bug.cgi?id=663992