Send key events to listeners assumed to be hung, but don't block
Orca would sometimes fail to notice that its modifier key was released
since it was presumed to be hung and would not be sent the event. So,
rather than ignore event listeners assumed to be hung, we'll send them
the notifications for the keys that they're listening to but not wait
for a reply, instead assuming that they are not consumed.