Reduce the method call timeout slightly
It is likely a good idea for the method call timeout to be lower than
the registry daemon's keystroke listener timeout. This way, if an AT
makes a method call that times out while in a keystroke event listener,
it will still have time to exit the listener before the registry daemon
presumes it hung.