[registry] Send a ping to a process when marking as hung
Sometimes keys that should be captured by event listeners are instead
passed through to the application. It seems that at-spi2-registryd might
not always be removing the process from the list of hung processes. I
have not reproduced this recently, and I'm not sure where the root
problem is. Hoping that it will help to send the process a ping and use
the reply as a queue that it is un-hung (previously we were relying on
receiving a reply to the original key event notification).

https://bugzilla.gnome.org/show_bug.cgi?id=671976