Increase the keystroke event timeout
It seems that it is currently possible for at-spi2-atk to get bogged
down when sending a keystroke notification, so 1 second is a short time
to wait for a reply. It may be possible to mitigate this in the future,
in which case we could reduce the timeout again, but not going to do
this for 3.4. See comment on
https://bugzilla.gnome.org/show_bug.cgi?id=671976.