Don't create a sub-directory for the socket when running as root
If an application is running as root but within a user's runtime directory,
then creating a temporary directory will result in the directory being owned
by root, and the user will not have permission to search the directory to
access the socket, so an AT running as the normal user will not be able to
connect to the application running as root.
Fixes regression introduced by the fix for BGO#678348.