Only create a directory and a socket when requested
Currently, atk-bridge clutters XDG_RUNTIME_DIR with directories for sockets,
which is especially annoying since gtk currently does not call
atk_bridge_adaptor_cleanup, so the directories never go away. This change does
not really solve the problem--it only hides it from users who do not really
need AT-SPI--but, nevertheless, there is no reason to create a directory or a
socket if nothing has requested it, so doing this lazily makes sense.