Only add items to the cache on children-changed if they are passed in
Gtk tree views, for instance, send children-changed signals without passing
an object. If a toolkit does this, then assume that it is doing so because
fetching the object would be expensive, and/or the object is transient, so
just have the cache bypass the signal rather than ref the child in order to
add it. Helps with performance when creating large gtk+ tree views.

https://bugzilla.gnome.org/show_bug.cgi?id=674719