Only send property-change signals relevant for caching, absent listeners
Previously, all property-change signals were passed on, regardless of any
registered listeners, because some of them are used by libatspi to keep the
cache up-to-date. However, gtk sends many notifications for property changes
that would have no impact on libatspi's caching code, so we should not waste
time building and sending D-Bus messages for them unless something is
interested in the event.

https://bugzilla.gnome.org/show_bug.cgi?id=674719