Only initialize the cache when an AT is running
If no AT is running, we shouldn't need to be maintaining the cache, so
only initialize when an AT is active. Note that this change does not
deinitialize the cache when an AT is no longer running. Deinitializing
and setting spi_global_cache to NULL might be a good idea, but trying to
keep risk to a minimum this close to 3.6.