Fix droute test, and remove dbind dependency
The droute test was broken (it had not been updated for recent API
changes and also did not return a non-zero exit code on failure). Also,
it is the only thing in at-spi2-atk that uses dbind, so re-worked it to
call libdbus directly, so that we can remove dbind / avoid having it
both here and in at-spi2-core.