Enable the building of introspection files for Visual C++ builds
This adds the utility Python/.bat scripts to the distribution for building
the introspection files for ATK, and adds utility projects to build those
files when needed.

This makes use of Python Regex functionality so that we can read directly
from the autotools files to create the file list that is necessary to
build the introspection files, without needing to modify the autotools
files to create the file list.