Revert "Do not run at-spi-dbus-bus at initialization phase"
The bus launcher may be started after applications that use AT-SPI, such
as caribou, if it is started in the application phase. I'm unsure of the
rationale for the original commit. I'm also unsure if the .desktop file
is still needed, since dbus-daemon should automatically launch the bus
launcher on request, but would rather not eliminate it this close to
3.8.

https://bugzilla.gnome.org/show_bug.cgi?id=694984

This reverts commit 1089e7c3687a48bb7810c40d6c9ff37e7acd10a1.