Use atspi_is_initialized to decide whether to expose p2p connection
Previously, direct D-Bus connections were not exposed if AT_SPI_CLIENT
was set in the environment, since they are causing deadlock if the
application and the AT are in the same process. However, AT_SPI_CLIENT
was only being set by pyatspi, meaning that libatspi consumers not using
pyatspi (Javascript embedded in gnome-shell that uses the libatspi
introspection bindings directly, for instance) would still have problems
when attempting to inspect their own accessibles.

Also removed the code in event.c to disable main loop context switching
when AT_SPI_CLIENT was set. Modifying it to also check
atspi_is_initialized would cause deadlock with gnome-shell when Orca is
running. Hopefully that code is not needed anymore.