Cancel pending calls when giving up on them
When timing out without having received a reply from a pending call,
then
cancel the call so that it will be removed from the DBusConnection's
queue. This will hopefully prevent the call and the
SpiReentrantCallClosure from being leaked.

Note that this code is still behaving oddly (why isn't a notification
being sent by libdbus when the specified amount of time has ellapsed?)