Use generated enum type for AtkRole to get the role name
Simplifies adding a new role (so less error prone) as with
this change a new role just needs to be added to the enum.

The specific role name is generated from the enum. For example:
ATK_ROLE_PUSH_BUTTON would create "push button"

It is also possible to specify the name. For example:
 ATK_ROLE_ACCEL_LABEL,      /*<nick=accelerator-label>*/
 would create "accelerator label"

Notice the '-'. At the enum it is needed to use it. Internally
all the names are compacted and maintained in one array (that
now also includes the extra role names).

https://bugzilla.gnome.org/show_bug.cgi?id=705290