build: Use Python to Generate the MSVC Projects
This would move the generation of the ATK Visual C++ 2008/2010 project
to Python 2/3 scripts from autotools scripts.  This would have the
following advantages:

-Reduce congestion in the autotools files, most notably atk/Makefile.am,
 and make everything that concerns the completion of MSVC project files
 fo under build/
-Easier to maintain and test, as a standard installation of Python (even
 on Windows) is enough to generate the Project files, and this can still
 be easily called during 'make dist'.
-Also paves the first steps for people wanting to build ATK from a GIT
 checkout, as this will help simplify the process

There is now a dependency on Python 2/3 for people that are wishing to do
'make dist', as naturally the scripts to do the Visual C++ project
generation is done with Python, but since one is likely going to generate
the .gir files for ATK when doing 'make dist'/'make distcheck', this is
satisfied as the scripts used to generate the .gir files are Python 2.6+
scripts as well.

https://bugzilla.gnome.org/show_bug.cgi?id=690145