build: Add NMake Makefiles for the Tests
This adds a set of NMake Makefiles that is used to build the test programs,
so that one would be able to build the test programs against a Visual C++
build of ATK quickly and easily.  As with the Project files, the listing of
test programs is also done with the Python script, which can also be done
under Windows or called by "make dist" or so quite easily.

https://bugzilla.gnome.org/show_bug.cgi?id=690145