doc: documentation for AtkTableCell, AtkTable, deprecations and padding
Index based methods forced ATK implementations to expose
the cells are direct children of the table. Something that
was complex in the practice. In fact some implementation
were not doing it, making the index-based methods not
properly implemented. This became even more clear with the
addition of AtkTableCell.

Additionally, this patch documents AtkTableCell and that
those cells should implement the newly added AtkTableCell
interface.