Don't pass empty property array to RegisterEvent
Apparently, for some reason unknown to me, it is sometimes possible for
a jhbuild session to be running an at-spi2-registryd from the system
directory, rather than the jhbuild install, so it is useful to be
compatible with at-spi2-registryd <= 3.10. So, for now, if no properties
are being requested with an event, then let's leave off the (empty)
array.

Also modify at-spi2-registryd to accept the old form of RegisterEvent.

We may want to revert this eventually; it would be cleaner not to send
two different forms of RegisterEvent.