configure: use AC_CHECK_ALIGNOF() instead of custom DBIND_CHECK_ALIGNOF()
DBIND_CHECK_ALIGNOF() is a custom autoconf macro from at-spi2-core
that tests the alignment required on the current platform for a given
C type. Unfortunately, this macro uses AC_TRY_RUN(), which requires
running a test program on the target platform, which is fundamentally
incompatible with cross-compilation.

Therefore, this patch switches the configure.ac code to use
AC_CHECK_ALIGNOF() instead, which is available in the standard set of
autoconf macros, and is able to test the required alignment for C
types without using AC_TRY_RUN(). Consequently, it adjusts the
dbind/dbind-config.h.in definitions to use the definitions created by
the AC_CHECK_ALIGNOF() macros, which are not substituted by autoconf,
but directly available from config.h.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>