MSVC Builds: "Install" .pdb Files
"Install" the .pdb files for Visual Studio builds, to make debugging builds
easier, especially when debugging the GTK+ stack.

Also make the "installation" of .lib and .dll files more selective so that
we won't accidently copy DLLs and LIBs that are either not meant to be
copied or have other specific places that they need to go, when the
project files here are included as part of a grand solution file, such as
one that is used to build the whole GTK+ stack from scratch.