MSVC Builds: Generate the Introspection Build Commands
This uses the common autotools module that was added in the previous commit
to generate the command lines used to generate the .gir and compile the
.typelib files, as well as the file list that is to be consumed during the
generation of the .gir files.

This also makes the detectmsvc-msvc.mak and introspection-msvc.mak NMake
modules the same as the ones in $(glib_srcroot)/build/win32 and
$(gi_srcroot)/build/win32 respectively, so that they may be copied from
these locations when they are updated.

https://bugzilla.gnome.org/show_bug.cgi?id=764983