Always accept connections as root if unable to find an owning user
When running as root, we look through /proc to try to identify our
parent and check that it matches the user making a connection. However,
this check does not always yield useful information and also will
always fail on anything that isn't Linux. Unsure what security
implications this change has, but then it isn't ideal to run UI as
the super-user in the first place, yet users sometimes need to do so
for various reasons, and a11y needs to be enabled, so adding this
change unless a better solution is found.