atkregistry: remove code not being compiled
I think that code was there in order to explain why it is not
needed. If that case, it would be better to just move the comment
around.

In general having code around #if 0 is a maintenance burden. That code
has not been compiled for years and nothing happen. It is better to
just remove it.