Make sure the runtime directory exists
It may happen that the home directory of the user does not contain the
runtime directory yet, when it's a freshly-created home for instance.
In that case, establishing the p2p socket in spi_atk_create_socket
will fail since it only tries to create a socket in that directory
without checking that the directory exists.

https://bugzilla.gnome.org/show_bug.cgi?id=763274