Correctly handle at-spi2-registryd dying when processing NameOwnerChanged
If at-spi2-registryd dies, then we receive two NameOwnerChanged signals,
one for org.a11y.atspi.Registry and the other for the initial dbus-issued
name. When we received the latter, we were matching it to the AtspiApplication
corresponding to the desktop and inadvertently disposing it, leading to a
crash later on.