docs: change id of the atkobject chapter
The reason is avoid collision with AtkObject.html on case-insensitive
filesystems.

One could argue that having a small html that just list AtkObject,
but this is done automatically by gtk-doc.

One could argue again, saying that probably the current chapter-split
main page (at atk-docs.sgml) is not the correct one, but I think
that for now a small id rename is better that a full main page
overhaul.

https://bugzilla.gnome.org/show_bug.cgi?id=772236