Visual Studio builds: Support Visual Studio 2017
Update the autotools scripts so that we can copy the Visual Studio 2010
project files and update them to obtain the Visual Studio 2017 projects.

As the format of the toolset version is different, allow a version number
string for the toolset version and use it if specified, otherwise just
create the toolset version string as we did before.

Note that Visual Studio 2015 and 2017 aims to be compatible in terms of
CRT usage, so it should be possible to use 2015-built binaries with
2017-built binaries.