Poll direct dbus connections in the main loop
If a process was marked as hung, then we send a ping with the intention
of removing it from the list of hung processes when a reply to the ping is
received. However, we weren't necessarily polling the connection, so we
wouldn't see the reply to the ping, meaning that the process could be
marked as hung indefinitely.

https://bugzilla.mozilla.org/show_bug.cgi?id=789038