Visual Studio builds: Integrate introspection builds
Since we hard-require Python for our builds, integrate introspection
builds in the build process.  This will fail if Python is not installed
or correctly set at PythonDir/PythonDirX64 or g-ir-scanner is not found
at $(GLibEtcInstallRoot)\bin, but it shouldn't otherwise impact the
build in other ways as the built DLL and lib will still be built and
copied normally.  This is in place here before the Meson build system is
ready for G-I and GLib.