Visual Studio builds: Don't build introspection by default
We aren't able to do a sanity check for g-ir-scanner in the NMake
Makefiles without extra tools, so it turns out that it would make it
more clear to people if we do not build the introspection files by
default here (especially that this is optional).

So, by doing so people could get the items needed for introspection
ready instead of build failures they do not expect, so people have
to build the introspect project explicitly, which will copy the built
introspection files to the appropriate location under
$(GlibEtcInstallPrefix) when sucessfully done.

Also fix the Visual Studio 2008 property sheets as a letter was
accidentaly added, breaking the builds.