support building a static library
Change the atspi library definition in meson.build to `library()`
instead of `shared_library()` so that when meson is called with
`--default-library static` a static library is built.

For this to work, also change the `declare_dependency()` in
`dbind/meson.build` to be a list of sources instead of a `link_with:`
`static_library()`.

https://bugzilla.gnome.org/show_bug.cgi?id=793762