bus-launch: add dbus-broker support
Both dbus-daemon and dbus-broker are now optional at compile-time, though
at least one must be configured. A new configuration option is introduce in
order to select the default implementation attempted at runtime. The other
implementation will function as a fall-back (in case support for both are
compiled in). If no default is selected, dbus-daemon remains the default as
before.

Unlike dbus-daemon, dbus-broker requires at-spi-bus-launch to create the
listening socket and pass it in, rather than having the bus do that and send
back the address. For now we follow what dbus-daemon does, and create a socket
in the abstract namespace, though it might be more suitable to create a socket
in $XDG_RUNTIME_DIR.

The only difference users should observe is that daemons are no longer spawned
by the bus implementation, but spawned and managed by the systemd user instance,
though this should not lead to a difference in behavior. In particular this
applies to `org.a11y.atspi.Registry`.

For non-linux and non-systemd systems, dbus-daemon should continue to be used.

[v2:
   - drop the --verbose switch, which is no longer supported
   - make dbus-daemon optional too
   - allow the default implementation to be selected]
Signed-off-by: Tom Gundersen <teg@jklm.no>