Do not pass unset variables to emit_event()
Uninitialized variables in C have undefined content; some compilers,
with some debugging flags, may decide to set them to 0, but you cannot
ever rely on that.

If the text variable is unset the validation when appending it to the
message payload may still pass, and we may end up sending random junk
over the wire — or, segfault.