docs: Remove atk.types
We can autogenerate it from the list of headers, and avoid it going out
of sync whenever we add a new GType.

In point of fact, we're already asking Meson/gtk-doc to do that, but if
we leave atk.types in the source directory, we're going to confuse Meson
and ourselves.