Clean up the documentation
Modify the sections file to ensure that there are no unused symbols.
This ensures that previously undocumented parts of the public API are
now in the documentation. Also move private symbols to Private
subsections and GObject stuff to Standard subsections. This ensures that
they don't show up in the documentation at all.

There is still plenty of work to get to a clean documentation build.