bus-launch: Do not poke X11 to check at-spi-bus-launcher is running
The already_running() check first gets the AT_SPI_BUS root window property,
then tries to open it to check if it exists. For it to exist though there
must be another at-spi-bus-launcher process around with the org.a11y.Bus
name.

It seems we can just defer the uniqueness check to g_bus_own_name(), as
we will get the "name lost" callback right after failing to acquire the
unique name. Doing it this way works for both x11 and non-x11, and avoids
assumptions on what the current windowing actually is.