Do not rely on getenv("DISPLAY")!=NULL to assume it is a X11 environment
Change/add checks around the AT_SPI_BUS root window property handling so it
is only done if the session is a real X11 one.

These checks used to work on wayland sessions, as there still is a X server
to poke, it's strange to use as a side channel but that's about it. However
in the future mutter will start Xwayland on demand, the DISPLAY environment
variable will definitely exist so checking for it is definitely not
sufficient, and opening the display will unintendedly spawn Xwayland.

It is debatable that this should happen in Wayland sessions at all, so let
the org.a11y.Bus fallbacks take over.